+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.1
+// swift-tools-version:5.3
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -14,8 +14,8 @@ let package = Package(
   dependencies: [
     // Dependencies declare other packages that this package depends on.
     // .package(url: /* package url */, from: "1.0.0"),
-    .package(url: "https://github.com/cfilipov/TextTable.git", .branch("master")),
-    .package(url: "https://github.com/apple/swift-argument-parser.git", .branch("master"))
+    .package(url: "https://github.com/thecb4/TextTable.git", .branch("master")),
+    .package(url: "https://github.com/apple/swift-argument-parser.git", .exact("0.3.1"))
   ],
   targets: [
     // Targets are the basic building blocks of a package. A target can define a module or a test suite.
@@ -28,10 +28,16 @@ let package = Package(
       dependencies: ["BlanketKit"]),
     .target(
       name: "BlanketCLI",
-      dependencies: ["BlanketKit", "ArgumentParser"]),
+      dependencies: [
+        "BlanketKit",
+        .product(name: "ArgumentParser", package: "swift-argument-parser")
+      ]),
     .testTarget(
       name: "BlanketCLITests",
-      dependencies: ["BlanketCLI", "ArgumentParser"]),
+      dependencies: [
+        "BlanketCLI",
+        .product(name: "ArgumentParser", package: "swift-argument-parser"),
+      ]),
     .target(
       name: "Blanket",
       dependencies: ["BlanketCLI"]),