+++ Tests/LinuxMain.swift
@@ -1,12 +1,1 @@
-import XCTest
-
-import BlanketCLITests
-import BlanketKitTests
-import BlanketTests
-
-var tests = [XCTestCaseEntry]()
-tests += BlanketCLITests.__allTests()
-tests += BlanketKitTests.__allTests()
-tests += BlanketTests.__allTests()
-
-XCTMain(tests)
+fatalError("Use `swift test --enable-test-discovery` to run tests")