+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 * Format based on [Keep A Change Log](https://keepachangelog.com/en/1.0.0/)
 * This project adheres to [Semantic Versioning](http://semver.org/). 
 
+## [Unreleased] - 2020-Dec-23
+### Changed
+- swift build tools to 5.3
+- swift-argument-parser set to 0.1.3 from master
+- text-table is now a fork from original
+- LinuxMain.swift in test now throws fatal error. Use `swift test --enable-test-discovery`
+- from `master` to `main` branch
+
+### Removed
+- test manifest files
+
+
 ## [0.1.0] - 2020-Feb-29 
 ### Added
 - default source folder